Twenty-sixth release (v0.28)
Another big release, that one come with fixes of many broken platforms and improvements on the webextension. ArtStation is back with a much better code! Twitter API URL changes has been included and since I made a DeviantArt account I polished it's support and can now download a higher resolution when the download button is available.

The desktop-app also received many useful bug fixes and now detect pixelart so it won't blur these when zomming in. I also managed to restore the support for Ubuntu 20.04 and made an experimental support for FlatPak, Arcollect does build and run in FlatPak but their bug on browsers nativeMessaging is blocking Arcollect and many other projects for now.